Refactor architecture with FastAPI scaffold and hardened dark theme CSS
